lz4_test.go: Fix/rename BenchmarkStreamUncompress (#24)
This benchmark previously failed to execute. The problem is it used
NewCompressReader to compress data, then tried to decompress it with
NewReader. These functions are not compatible. Fix the benchmark to
use NewWriter() to compress and NewReader() to decompress. Rename
the benchmark to Deprecated, since the NewReader() function is marked
as deprecated.

The NewCompressReader/NewDecompressReader pair are benchmarked in
BenchmarkStreamDecompressReader.